+++ test/python/test_circuit.py
@@ -5,15 +5,89 @@
 # This source code is licensed under the Apache License, Version 2.0 found in
 # the LICENSE.txt file in the root directory of this source tree.
 
-# pylint: disable=unused-import
+# pylint: disable=unused-import,invalid-name
 
-"""Tests for combining and extending circuits across width and depth"""
+"""Test Qiskit's QuantumCircuit class."""
 
 import qiskit.extensions.simulator
-from qiskit import (ClassicalRegister, QISKitError, QuantumCircuit,
-                    QuantumRegister, execute)
+from qiskit import QuantumRegister, ClassicalRegister, QuantumCircuit
+from qiskit import execute
+from qiskit import QISKitError
 from qiskit.tools.qi.qi import state_fidelity
 from .common import QiskitTestCase
+
+
+class TestCircuit(QiskitTestCase):
+    """QuantumCircuit basic tests."""
+
+    def test_get_qregs(self):
+        """Test getting quantum registers from circuit.
+        """
+        qr1 = QuantumRegister(1)
+        qr2 = QuantumRegister(2)
+        qc = QuantumCircuit(qr1, qr2)
+        q_regs = qc.get_qregs()
+        self.assertEqual(len(q_regs), 2)
+        self.assertEqual(q_regs[qr1.name], qr1)
+        self.assertEqual(q_regs[qr2.name], qr2)
+
+    def test_get_cregs(self):
+        """Test getting classical registers from circuit.
+        """
+        cr1 = ClassicalRegister(1)
+        cr2 = ClassicalRegister(2)
+        cr3 = ClassicalRegister(3)
+        qc = QuantumCircuit(cr1, cr2, cr3)
+        c_regs = qc.get_cregs()
+        self.assertEqual(len(c_regs), 3)
+        self.assertEqual(c_regs[cr1.name], cr1)
+        self.assertEqual(c_regs[cr2.name], cr2)
+
+    def test_circuit_qasm(self):
+        """Test circuit qasm() method.
+        """
+        qr1 = QuantumRegister(1, 'qr1')
+        qr2 = QuantumRegister(2, 'qr2')
+        c = ClassicalRegister(3, 'c')
+        qc = QuantumCircuit(qr1, qr2, c)
+        qc.u1(0.3, qr1[0])
+        qc.u2(0.2, 0.1, qr2[0])
+        qc.u3(0.3, 0.2, 0.1, qr2[1])
+        qc.s(qr2[1])
+        qc.s(qr2[1]).inverse()
+        qc.cx(qr1[0], qr2[1])
+        qc.barrier(qr2)
+        qc.cx(qr2[1], qr1[0])
+        qc.h(qr2[1])
+        qc.x(qr2[1]).c_if(c, 0)
+        qc.y(qr1[0]).c_if(c, 1)
+        qc.z(qr1[0]).c_if(c, 2)
+        qc.barrier(qr1, qr2)
+        qc.measure(qr1[0], c[0])
+        qc.measure(qr2[0], c[1])
+        qc.measure(qr2[1], c[2])
+        expected_qasm = """OPENQASM 2.0;
+include "qelib1.inc";
+qreg qr1[1];
+qreg qr2[2];
+creg c[3];
+u1(0.300000000000000) qr1[0];
+u2(0.200000000000000,0.100000000000000) qr2[0];
+u3(0.300000000000000,0.200000000000000,0.100000000000000) qr2[1];
+s qr2[1];
+sdg qr2[1];
+cx qr1[0],qr2[1];
+barrier qr2[0],qr2[1];
+cx qr2[1],qr1[0];
+h qr2[1];
+if(c==0) x qr2[1];
+if(c==1) y qr1[0];
+if(c==2) z qr1[0];
+barrier qr1[0],qr2[0],qr2[1];
+measure qr1[0] -> c[0];
+measure qr2[0] -> c[1];
+measure qr2[1] -> c[2];\n"""
+        self.assertEqual(qc.qasm(), expected_qasm)
 
 
 class TestCircuitCombineExtend(QiskitTestCase):
@@ -66,10 +140,10 @@ class TestCircuitCombineExtend(QiskitTestCase):
         cr1 = ClassicalRegister(1, "q")
         qc1 = QuantumCircuit(qr1)
         qc2 = QuantumCircuit(qr2)
-        qc3 = QuantumCircuit(cr1)
+        qcr3 = QuantumCircuit(cr1)
 
         self.assertRaises(QISKitError, qc1.__add__, qc2)
-        self.assertRaises(QISKitError, qc1.__add__, qc3)
+        self.assertRaises(QISKitError, qc1.__add__, qcr3)
 
     def test_combine_circuit_extension_instructions(self):
         """Test combining circuits contining barrier, initializer, snapshot
@@ -144,10 +218,10 @@ class TestCircuitCombineExtend(QiskitTestCase):
         cr1 = ClassicalRegister(1, "q")
         qc1 = QuantumCircuit(qr1)
         qc2 = QuantumCircuit(qr2)
-        qc3 = QuantumCircuit(cr1)
+        qcr3 = QuantumCircuit(cr1)
 
         self.assertRaises(QISKitError, qc1.__iadd__, qc2)
-        self.assertRaises(QISKitError, qc1.__iadd__, qc3)
+        self.assertRaises(QISKitError, qc1.__iadd__, qcr3)
 
     def test_extend_circuit_extension_instructions(self):
         """Test extending circuits contining barrier, initializer, snapshot